Text.Hyphenation.Exception

Description

Pattern file support

data Exceptions

Hyphenation exceptions are special cases that should use the specified hyphenation points.

addException :: String -> Exceptions -> Exceptions

add an exception to the exception table. if it is already present, this will restrict the set of hyphenations to the intersection of the set provided and the set present.

lookupException :: String -> Exceptions -> Maybe [Int]

Try to find a matching hyphenation exception.

scoreException :: String -> [Int]

Convert an exception string to a score.

parseExceptions :: String -> Exceptions

Parse one exception per line from an input string