Text.Hyphenation.Language

Description

Pattern file support

data Language

A strongly typed set of available languages you can use for hyphenation.

Constructors

Afrikaans

Basque

Bengali

Bulgarian

Catalan

Chinese

Coptic

Croatian

Czech

Danish

Dutch

English_US

English_GB

Esperanto

Estonian

Ethiopic

Finnish
Farsi

French

Friulan

Galician

Georgian

German_1901

German_1996

German_Swiss

Greek_Ancient

Greek_Mono

Greek_Poly

Gujarati

Hindi

Hungarian

Icelandic

Indonesian

Interlingua

Irish

Italian

Kannada

Kurmanji

Lao

Latin

Latin_Classic

Latvian

Lithuanian

Malayalam

Marathi

Mongolian

Norwegian_Bokmal

Norwegian_Nynorsk

Oriya

Panjabi

Piedmontese

Polish

Portuguese

Romanian

Romansh

Russian

Sanskrit

Serbian_Cyrillic

Serbocroatian_Cyrillic

Serbocroatian_Latin

Slovak

Slovenian

Spanish

Swedish

Tamil

Telugu

Thai

Turkish

Turkmen

Ukrainian

Uppersorbian

Welsh

languageHyphenator :: Language -> Hyphenator

Load (and cache) the hyphenator for a given language.

Provided language hyphenators

afrikaans :: Hyphenator

Hyphenators for a wide array of languages.

basque :: Hyphenator

Hyphenators for a wide array of languages.

bengali :: Hyphenator

Hyphenators for a wide array of languages.

bulgarian :: Hyphenator

Hyphenators for a wide array of languages.

catalan :: Hyphenator

Hyphenators for a wide array of languages.

chinese :: Hyphenator

Hyphenators for a wide array of languages.

coptic :: Hyphenator

Hyphenators for a wide array of languages.

croatian :: Hyphenator

Hyphenators for a wide array of languages.

czech :: Hyphenator

Hyphenators for a wide array of languages.

danish :: Hyphenator

Hyphenators for a wide array of languages.

dutch :: Hyphenator

Hyphenators for a wide array of languages.

english_US :: Hyphenator

>>> hyphenate english_US "supercalifragilisticexpialadocious"
["su","per","cal","ifrag","ilis","tic","ex","pi","al","ado","cious"]

favors US hyphenation

english_GB :: Hyphenator

>>> hyphenate english_GB "supercalifragilisticexpialadocious"
["su","per","cal","i","fra","gil","istic","ex","pi","alado","cious"]

favors UK hyphenation

esperanto :: Hyphenator

Hyphenators for a wide array of languages.

estonian :: Hyphenator

Hyphenators for a wide array of languages.

ethiopic :: Hyphenator

Hyphenators for a wide array of languages.

finnish :: Hyphenator

Hyphenators for a wide array of languages.

french :: Hyphenator

>>> hyphenate french "anticonstitutionnellement"
["an","ti","cons","ti","tu","tion","nel","le","ment"]

friulan :: Hyphenator

Hyphenators for a wide array of languages.

galician :: Hyphenator

Hyphenators for a wide array of languages.

georgian :: Hyphenator

Hyphenators for a wide array of languages.

german_1901 :: Hyphenator

Hyphenators for a wide array of languages.

german_1996 :: Hyphenator

Hyphenators for a wide array of languages.

german_Swiss :: Hyphenator

Hyphenators for a wide array of languages.

greek_Ancient :: Hyphenator

Hyphenators for a wide array of languages.

greek_Mono :: Hyphenator

Hyphenators for a wide array of languages.

greek_Poly :: Hyphenator

Hyphenators for a wide array of languages.

gujarati :: Hyphenator

Hyphenators for a wide array of languages.

hindi :: Hyphenator

Hyphenators for a wide array of languages.

hungarian :: Hyphenator

Hyphenators for a wide array of languages.

icelandic :: Hyphenator

>>> hyphenate icelandic "va\240lahei\240avegavinnuverkf\230rageymslusk\250r"
["va\240la","hei\240a","vega","vinnu","verk","f\230ra","geymslu","sk\250r"]

indonesian :: Hyphenator

Hyphenators for a wide array of languages.

interlingua :: Hyphenator

Hyphenators for a wide array of languages.

irish :: Hyphenator

Hyphenators for a wide array of languages.

italian :: Hyphenator

Hyphenators for a wide array of languages.

kannada :: Hyphenator

Hyphenators for a wide array of languages.

kurmanji :: Hyphenator

Hyphenators for a wide array of languages.

lao :: Hyphenator

Hyphenators for a wide array of languages.

latin :: Hyphenator

Hyphenators for a wide array of languages.

latin_Classic :: Hyphenator

Hyphenators for a wide array of languages.

latvian :: Hyphenator

Hyphenators for a wide array of languages.

lithuanian :: Hyphenator

Hyphenators for a wide array of languages.

malayalam :: Hyphenator

Hyphenators for a wide array of languages.

marathi :: Hyphenator

Hyphenators for a wide array of languages.

mongolian :: Hyphenator

Hyphenators for a wide array of languages.

norwegian_Bokmal :: Hyphenator

Hyphenators for a wide array of languages.

norwegian_Nynorsk :: Hyphenator

Hyphenators for a wide array of languages.

oriya :: Hyphenator

Hyphenators for a wide array of languages.

panjabi :: Hyphenator

Hyphenators for a wide array of languages.

piedmontese :: Hyphenator

Hyphenators for a wide array of languages.

polish :: Hyphenator

Hyphenators for a wide array of languages.

portuguese :: Hyphenator

Hyphenators for a wide array of languages.

romanian :: Hyphenator

Hyphenators for a wide array of languages.

romansh :: Hyphenator

Hyphenators for a wide array of languages.

russian :: Hyphenator

Hyphenators for a wide array of languages.

sanskrit :: Hyphenator

Hyphenators for a wide array of languages.

serbian_Cyrillic :: Hyphenator

Hyphenators for a wide array of languages.

serbocroatian_Cyrillic :: Hyphenator

Hyphenators for a wide array of languages.

serbocroatian_Latin :: Hyphenator

Hyphenators for a wide array of languages.

slovak :: Hyphenator

Hyphenators for a wide array of languages.

slovenian :: Hyphenator

Hyphenators for a wide array of languages.

spanish :: Hyphenator

Hyphenators for a wide array of languages.

swedish :: Hyphenator

Hyphenators for a wide array of languages.

tamil :: Hyphenator

Hyphenators for a wide array of languages.

telugu :: Hyphenator

Hyphenators for a wide array of languages.

thai :: Hyphenator

Hyphenators for a wide array of languages.

turkish :: Hyphenator

Hyphenators for a wide array of languages.

turkmen :: Hyphenator

Hyphenators for a wide array of languages.

ukrainian :: Hyphenator

Hyphenators for a wide array of languages.

uppersorbian :: Hyphenator

Hyphenators for a wide array of languages.

welsh :: Hyphenator

Hyphenators for a wide array of languages.

loadHyphenator :: String -> IO Hyphenator

Read a built-in language file from the data directory where cabal installed this package.

(e.g. hyphenateLanguage "en-us" opens "/Users/ekmett/.cabal/share/hyphenation-0.2/ghc-7.4.1/hyph-en-us.hyp.txt" among others when run on the author's local machine)

languageAffix :: Language -> String

the infix portion of the data file names used for this language