Text.Hyphenation.Pattern

Description

Pattern file support

data Patterns

Hyphenation patterns

insertPattern :: String -> Patterns -> Patterns

Insert a Knuth-Liang hyphenation pattern into the trie

. denotes the start or end of the input

0-9 are used to denote hyphenation or dehyphenation depending on whether or not they are even (no hyphen) or odd (hyphen allowed).

Patterns are overlaid and the maximum value at each location is used. this allows you to implement a finite number of precedences between hyphenation rules

(e.g. 3foo. indicates that the suffix '-foo' should be hyphenated with precedence 3.)

lookupPattern :: String -> Patterns -> [Int]

Tallies the hyphenation scores for a word considering all tails.

scorePattern :: String -> [Int]

Convert a Pattern string to a Score

parsePatterns :: String -> Patterns

Parse one pattern per line from an input string

hyph-utf8 supplies these files UTF-8 encoded in the txt folder with a .pat.txt extension